Fix geometry with attributes.
When the property 'attributes' was specified the code was looking
in there for materialProperties, now it looks in both places.
This was causing geometry with attributes to not render and
some times error. They should display fine now.
Added regression test.
These changes are applied to genie in https://cr.flux.io/#/c/9686/

Change-Id: I1c558490889565a73a0e013070405bad39862326