Allow objects with equivalent materials to merge - LIB3D-5
Having many objects in the rendering scene is very slow. Thus we merge when
we can. Previously we only merged if there was no material, but because
of a default material being applied, the merge almost never happened.
Now we check if the materials are the same, and merge when appropriate.
Also integrated logic for reflective materials with roughness.
Backs https://cr.flux.io/10660

Change-Id: I725d43cdbe702d16ced708b63be7f4840f5554e7