LIB3D-819: Revit Element display in viewport.
In this CL I am adding support to display revit element that
contains a mesh in the web viewer.

:maker-facing:

QA: Make sure that a single revit elment that contains a mesh should automatically
activate the 3D view in the toolbar. Revit element with mesh can be obtained
using the latest dev version of the revit plugin. When switching to that
view the elements mesh should be displayed. When data inspector contains
an array of revit elements, any and all elements that contain a mesh should be displayed in the
viewport. Display mesh is located at object.geometryParameters.geometry
in the revit element. When more than one revit elements are overlayed on a
data inspector in the Data view, they should display with different materials.

Change-Id: I83540d27be5b81ffa04fb3487f93a187fbfa0a49