LIB3D-901: Flux-viewport switches to 3D view even if revit element does not have geometry.
In this CL I am making sure that the flux-viewport does not switch to 3D view
when none of the revit elements in key do not have a geometry.

:maker-facing:

QA:
1) Make sure that single element with revitElement that contains geometry
(geometryParameters/geometry NOT equal to null) the view port switches to
3D View.
2) Make sure that with array of revit element even if one object has geometry
the viewport switches to 3D View.
4) Make sure that single revit element with  no geometry the viewport does not
switch to 3d View
5) Make sure that with array of revit elements if none of the elements have
geoemtry, viewport does not switch to 3D view.

Change-Id: I10d8570bdd68374cd5763449342388d67ab0124c