Fill in missing catch for certain errors - LIB3D-1100
In some situations a timeout would cause an exception in a dependency
of the viewport. This was not caught, causing the viewport to free.
Now this is handled with the promise except clause.

:hp: :viewport:

Code Review:
1. Check out related CLs.
2. make run-viewport
QA:
You need to create a brep that times out when rendering.
The way I ran into that was the following:
0. Open the browser console
1. Create the Modeling Demo app
2. Scroll down to the loft stream
3. Create a scale block and scale up the loft model by a few times.
4. Render that in a viewport (see attached images).
5. It should time out and have the error message, not be stuck loading.
I cant guarantee this will work, but it has consistently for me.
See images for right and wrong behavior.
Change-Id: Icd7ff18cce5163078a2d970d8766f3a0be02205e